[FEAT] 사용자 정보 API 구현 및 jwt 만료기한 설정 수정

+++ src/main/java/com/lucky/around/meal/entity/Member.java
@@ -32,8 +32,8 @@ public class Member {
   private MemberRole role;
 
   @Column(nullable = true)
-  private long lat;
+  private double lat;
 
   @Column(nullable = true)
-  private long lon;
+  private double lon;
 }
